requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langchain
langchain-groq
